尝试修复电机接入后编码器异常问题：Motor_SetSpeed 参数改为 int16_t

+++ Hardware/Motor.c
@@ -16,7 +16,7 @@ void Motor_Init(void)
 	
 }
 
-void Motor_SetSpeed(int8_t Speed)
+void Motor_SetSpeed(int16_t Speed)
 {
 	if (Speed > 0)
 	{
